requirements.txt
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-exporter-otlp-proto-grpc==1.33.0
